requirements.txt
fastapi
uvicorn
gpt4all
python-dotenv
pandas
scikit-learn
joblib
